split receiver into per-functionality files
