+++ receiver/factory_test.go
@@ -1,0 +1,93 @@
+// Copyright The OpenTelemetry Authors
+// SPDX-License-Identifier: Apache-2.0
+
+package receiver
+
+import (
+	"context"
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+
+	"go.opentelemetry.io/collector/component"
+	"go.opentelemetry.io/collector/consumer/consumertest"
+)
+
+func TestNewFactory(t *testing.T) {
+	var testType = component.MustNewType("test")
+	defaultCfg := struct{}{}
+	factory := NewFactory(
+		testType,
+		func() component.Config { return &defaultCfg })
+	assert.EqualValues(t, testType, factory.Type())
+	assert.EqualValues(t, &defaultCfg, factory.CreateDefaultConfig())
+	_, err := factory.CreateTracesReceiver(context.Background(), CreateSettings{}, &defaultCfg, consumertest.NewNop())
+	assert.Error(t, err)
+	_, err = factory.CreateMetricsReceiver(context.Background(), CreateSettings{}, &defaultCfg, consumertest.NewNop())
+	assert.Error(t, err)
+	_, err = factory.CreateLogsReceiver(context.Background(), CreateSettings{}, &defaultCfg, consumertest.NewNop())
+	assert.Error(t, err)
+}
+
+func TestNewFactoryWithOptions(t *testing.T) {
+	var testType = component.MustNewType("test")
+	defaultCfg := struct{}{}
+	factory := NewFactory(
+		testType,
+		func() component.Config { return &defaultCfg },
+		WithTraces(createTraces, component.StabilityLevelDeprecated),
+		WithMetrics(createMetrics, component.StabilityLevelAlpha),
+		WithLogs(createLogs, component.StabilityLevelStable))
+	assert.EqualValues(t, testType, factory.Type())
+	assert.EqualValues(t, &defaultCfg, factory.CreateDefaultConfig())
+
+	assert.Equal(t, component.StabilityLevelDeprecated, factory.TracesReceiverStability())
+	_, err := factory.CreateTracesReceiver(context.Background(), CreateSettings{}, &defaultCfg, nil)
+	assert.NoError(t, err)
+
+	assert.Equal(t, component.StabilityLevelAlpha, factory.MetricsReceiverStability())
+	_, err = factory.CreateMetricsReceiver(context.Background(), CreateSettings{}, &defaultCfg, nil)
+	assert.NoError(t, err)
+
+	assert.Equal(t, component.StabilityLevelStable, factory.LogsReceiverStability())
+	_, err = factory.CreateLogsReceiver(context.Background(), CreateSettings{}, &defaultCfg, nil)
+	assert.NoError(t, err)
+}
+
+func TestMakeFactoryMap(t *testing.T) {
+	type testCase struct {
+		name string
+		in   []Factory
+		out  map[component.Type]Factory
+	}
+
+	p1 := NewFactory(component.MustNewType("p1"), nil)
+	p2 := NewFactory(component.MustNewType("p2"), nil)
+	testCases := []testCase{
+		{
+			name: "different names",
+			in:   []Factory{p1, p2},
+			out: map[component.Type]Factory{
+				p1.Type(): p1,
+				p2.Type(): p2,
+			},
+		},
+		{
+			name: "same name",
+			in:   []Factory{p1, p2, NewFactory(component.MustNewType("p1"), nil)},
+		},
+	}
+
+	for i := range testCases {
+		tt := testCases[i]
+		t.Run(tt.name, func(t *testing.T) {
+			out, err := MakeFactoryMap(tt.in...)
+			if tt.out == nil {
+				assert.Error(t, err)
+				return
+			}
+			assert.NoError(t, err)
+			assert.Equal(t, tt.out, out)
+		})
+	}
+}
