+++ receiver/logs.go
@@ -1,0 +1,36 @@
+// Copyright The OpenTelemetry Authors
+// SPDX-License-Identifier: Apache-2.0
+
+package receiver // import "go.opentelemetry.io/collector/receiver"
+
+import (
+	"context"
+
+	"go.opentelemetry.io/collector/component" // Metrics receiver receives metrics.
+	"go.opentelemetry.io/collector/consumer"
+)
+
+// Logs receiver receives logs.
+// Its purpose is to translate data from any format to the collector's internal logs data format.
+// LogsReceiver feeds a consumer.Logs with data.
+//
+// For example, it could be a receiver that reads syslogs and convert them into plog.Logs.
+type Logs interface {
+	component.Component
+}
+
+// CreateLogsFunc is the equivalent of ReceiverFactory.CreateLogsReceiver().
+type CreateLogsFunc func(context.Context, CreateSettings, component.Config, consumer.Logs) (Logs, error)
+
+// CreateLogsReceiver implements Factory.CreateLogsReceiver().
+func (f CreateLogsFunc) CreateLogsReceiver(
+	ctx context.Context,
+	set CreateSettings,
+	cfg component.Config,
+	nextConsumer consumer.Logs,
+) (Logs, error) {
+	if f == nil {
+		return nil, component.ErrDataTypeIsNotSupported
+	}
+	return f(ctx, set, cfg, nextConsumer)
+}
