+++ receiver/metrics.go
@@ -1,0 +1,35 @@
+// Copyright The OpenTelemetry Authors
+// SPDX-License-Identifier: Apache-2.0
+
+package receiver // import "go.opentelemetry.io/collector/receiver"
+
+import (
+	"context"
+
+	"go.opentelemetry.io/collector/component" // Metrics receiver receives metrics.
+	"go.opentelemetry.io/collector/consumer"
+)
+
+// Its purpose is to translate data from any format to the collector's internal metrics format.
+// MetricsReceiver feeds a consumer.Metrics with data.
+//
+// For example, it could be Prometheus data source which translates Prometheus metrics into pmetric.Metrics.
+type Metrics interface {
+	component.Component
+}
+
+// CreateMetricsFunc is the equivalent of Factory.CreateMetrics.
+type CreateMetricsFunc func(context.Context, CreateSettings, component.Config, consumer.Metrics) (Metrics, error)
+
+// CreateMetricsReceiver implements Factory.CreateMetricsReceiver().
+func (f CreateMetricsFunc) CreateMetricsReceiver(
+	ctx context.Context,
+	set CreateSettings,
+	cfg component.Config,
+	nextConsumer consumer.Metrics,
+) (Metrics, error) {
+	if f == nil {
+		return nil, component.ErrDataTypeIsNotSupported
+	}
+	return f(ctx, set, cfg, nextConsumer)
+}
