+++ receiver/traces.go
@@ -1,0 +1,34 @@
+// Copyright The OpenTelemetry Authors
+// SPDX-License-Identifier: Apache-2.0
+
+package receiver // import "go.opentelemetry.io/collector/receiver"
+
+import (
+	"context"
+
+	"go.opentelemetry.io/collector/component" // Traces receiver receives traces.
+	"go.opentelemetry.io/collector/consumer"
+)
+
+// Its purpose is to translate data from any format to the collector's internal trace format.
+// TracesReceiver feeds a consumer.Traces with data.
+//
+// For example, it could be Zipkin data source which translates Zipkin spans into ptrace.Traces.
+type Traces interface {
+	component.Component
+}
+
+// CreateTracesFunc is the equivalent of Factory.CreateTraces.
+type CreateTracesFunc func(context.Context, CreateSettings, component.Config, consumer.Traces) (Traces, error)
+
+// CreateTracesReceiver implements Factory.CreateTracesReceiver().
+func (f CreateTracesFunc) CreateTracesReceiver(
+	ctx context.Context,
+	set CreateSettings,
+	cfg component.Config,
+	nextConsumer consumer.Traces) (Traces, error) {
+	if f == nil {
+		return nil, component.ErrDataTypeIsNotSupported
+	}
+	return f(ctx, set, cfg, nextConsumer)
+}
